requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aom=3.8.2=h63175ca_0
asttokens=2.4.1=pyhd8ed1ab_0
blas=1.0=mkl
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotli-python=1.0.9=py312hd77b12b_7
bzip2=1.0.8=h2bbff1b_5
ca-certificates=2024.3.11=haa95532_0
cairo=1.18.0=h1fef639_0
certifi=2024.2.2=py312haa95532_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.0=py312h59b6b97_0
cuda-cccl=12.4.127=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.127=0
cuda-opencl-dev=12.4.127=0
cuda-profiler-api=12.4.127=0
cuda-runtime=12.1.0=0
cuda-version=11.8=hcce14f8_2
cudatoolkit=11.8.0=hd77b12b_0
cudnn=8.9.7.29=he6de189_3
cycler=0.11.0=pyhd3eb1b0_0
dav1d=1.2.1=h2bbff1b_0
debugpy=1.8.1=py312h53d5487_0
decorator=5.1.1=pyhd8ed1ab_0
dlib=19.24.4=cuda118py312h0b77f89_0
double-conversion=3.3.0=h63175ca_0
easyprocess=1.1=pyhd8ed1ab_0
entrypoint2=1.1=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
expat=2.6.2=hd77b12b_0
ffmpeg=6.1.1=gpl_h66c0b5b_108
filelock=3.13.1=py312haa95532_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.14.2=hbde0cde_0
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
fonttools=4.51.0=py312h2bbff1b_0
freeglut=3.2.2=h63175ca_2
freetype=2.12.1=ha860e81_0
gettext=0.22.5=h5728263_2
gettext-tools=0.22.5=h7d00a51_2
glib=2.80.0=h39d0aa6_6
glib-tools=2.80.0=h0a98069_6
graphite2=1.3.14=hd77b12b_1
gst-plugins-base=1.24.1=h001b923_1
gstreamer=1.24.1=hb4038d2_1
harfbuzz=8.4.0=h7ab893a_0
hdf5=1.14.3=nompi_h73e8ff5_100
icc_rt=2022.1.0=h6049295_2
icu=73.2=h63175ca_0
idna=3.4=py312haa95532_0
imath=3.1.11=h12be248_0
importlib-metadata=7.1.0=pyha770c72_0
importlib_metadata=7.1.0=hd8ed1ab_0
imutils=0.5.4=py312h2e8e312_4
intel-openmp=2023.2.0=h57928b3_50497
ipykernel=6.29.3=pyha63f2e9_0
ipython=8.22.2=pyh7428d3b_0
jasper=4.2.3=h28f2b1a_0
jedi=0.19.1=pyhd8ed1ab_0
jeepney=0.7.1=pyhd3eb1b0_0
jinja2=3.1.3=py312haa95532_0
jupyter_client=8.6.1=pyhd8ed1ab_0
jupyter_core=5.7.2=py312h2e8e312_0
khronos-opencl-icd-loader=2023.04.17=h64bf75a_0
kiwisolver=1.4.4=py312hd77b12b_0
krb5=1.21.2=heb0366b_0
lcms2=2.16=h67d730c_0
lerc=4.0.0=h63175ca_0
libabseil=20240116.1=cxx17_h63175ca_2
libaec=1.1.3=h63175ca_0
libasprintf=0.22.5=h5728263_2
libasprintf-devel=0.22.5=h5728263_2
libblas=3.9.0=20_win64_mkl
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libcblas=3.9.0=20_win64_mkl
libclang13=18.1.3=default_hf64faad_0
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.5.147=0
libcurand-dev=10.3.5.147=0
libcurl=8.5.0=h86230a5_0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.20=hcfcfb64_0
libexpat=2.6.2=h63175ca_0
libffi=3.4.4=hd77b12b_0
libgettextpo=0.22.5=h5728263_2
libgettextpo-devel=0.22.5=h5728263_2
libglib=2.80.0=h39d0aa6_6
libhwloc=2.10.0=default_h2fffb23_1000
libiconv=1.17=hcfcfb64_2
libintl=0.22.5=h5728263_2
libintl-devel=0.22.5=h5728263_2
libjpeg-turbo=3.0.0=hcfcfb64_1
liblapack=3.9.0=20_win64_mkl
liblapacke=3.9.0=20_win64_mkl
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libogg=1.3.5=h2bbff1b_1
libopencv=4.9.0=qt6_py312hd35d245_612
libopenvino=2024.0.0=hfe1841e_5
libopenvino-auto-batch-plugin=2024.0.0=h04f32e0_5
libopenvino-auto-plugin=2024.0.0=h04f32e0_5
libopenvino-hetero-plugin=2024.0.0=h372dad0_5
libopenvino-intel-cpu-plugin=2024.0.0=hfe1841e_5
libopenvino-intel-gpu-plugin=2024.0.0=hfe1841e_5
libopenvino-ir-frontend=2024.0.0=h372dad0_5
libopenvino-onnx-frontend=2024.0.0=hdeef14f_5
libopenvino-paddle-frontend=2024.0.0=hdeef14f_5
libopenvino-pytorch-frontend=2024.0.0=he0c23c2_5
libopenvino-tensorflow-frontend=2024.0.0=h7c40eac_5
libopenvino-tensorflow-lite-frontend=2024.0.0=he0c23c2_5
libopus=1.3.1=h8ffe710_1
libpng=1.6.43=h19919ed_0
libprotobuf=4.25.3=h503648d_0
libsodium=1.0.18=h8d14728_1
libsqlite=3.45.3=hcfcfb64_0
libssh2=1.10.0=he2ea4bf_2
libtiff=4.6.0=hddb2be6_3
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=he774522_0
libwebp-base=1.3.2=h2bbff1b_0
libxcb=1.15=hcd874cb_0
libxml2=2.12.6=hc3477c8_2
libzlib=1.2.13=hcfcfb64_5
libzlib-wapi=1.2.13=hcfcfb64_5
lz4-c=1.9.4=h2bbff1b_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.3=py312h2bbff1b_0
matplotlib=3.8.4=py312h2e8e312_0
matplotlib-base=3.8.4=py312hc7c4135_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mkl=2023.2.0=h6a75c08_50497
mkl-service=2.4.0=py312h2bbff1b_1
mkl_fft=1.3.8=py312h2bbff1b_0
mkl_random=1.2.4=py312h59b6b97_0
mouseinfo=0.1.3=pyhd8ed1ab_0
mpmath=1.3.0=py312haa95532_0
msys2-conda-epoch=20160418=1
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.1=py312haa95532_0
numpy=1.26.4=py312hfd52020_0
numpy-base=1.26.4=py312h4dde369_0
opencv=4.9.0=qt6_py312hfab33ed_612
openexr=3.2.2=h72640d8_1
openh264=2.4.1=h63175ca_0
openjpeg=2.5.2=h3d672ee_0
openssl=3.2.1=hcfcfb64_1
packaging=24.0=pyhd8ed1ab_0
parso=0.8.4=pyhd8ed1ab_0
pcre2=10.43=h17e33f8_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=py312h6f6a607_0
pip=23.3.1=py312haa95532_0
pixman=0.43.4=h63175ca_0
platformdirs=4.2.1=pyhd8ed1ab_0
ply=3.11=py312haa95532_1
prompt-toolkit=3.0.42=pyha770c72_0
psutil=5.9.8=py312he70551f_0
pthread-stubs=0.3=h3c9f919_1
pthreads-win32=2.9.1=hfa6e2cd_3
pugixml=1.14=h63175ca_0
pure_eval=0.2.2=pyhd8ed1ab_0
py-opencv=4.9.0=qt6_py312hb11cbc4_612
pyautogui=0.9.54=py312h2e8e312_2
pybind11-abi=5=hd3eb1b0_0
pygetwindow=0.0.9=py312h2e8e312_2
pygments=2.17.2=pyhd8ed1ab_0
pymsgbox=1.0.9=pyhd8ed1ab_1
pyparsing=3.0.9=py312haa95532_0
pyperclip=1.8.2=pyhd8ed1ab_2
pyqt=5.15.10=py312hd77b12b_0
pyqt5-sip=12.13.0=py312h2bbff1b_0
pyrect=0.2.0=pyhd8ed1ab_0
pyscreenshot=3.1=pyhd8ed1ab_0
pyscreeze=0.1.30=pyhd8ed1ab_0
pysocks=1.7.1=py312haa95532_0
python=3.12.3=h2628c8c_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-mss=9.0.1=pyhd8ed1ab_0
python-xlib=0.33=pyhd8ed1ab_0
python_abi=3.12=4_cp312
pytorch=2.2.2=py3.12_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-mutex=1.0=cuda
pytweening=1.2.0=pyhd8ed1ab_0
pywin32=306=py312h53d5487_2
pyyaml=6.0.1=py312h2bbff1b_0
pyzmq=26.0.2=py312hd7027bb_0
qt-main=5.15.8=hcef0176_21
qt6-main=6.6.3=h5dee92f_0
requests=2.31.0=py312haa95532_1
rubicon-objc=0.4.8=pyhd8ed1ab_0
scipy=1.13.0=py312hbb039d4_0
setuptools=68.2.2=py312haa95532_0
sip=6.7.12=py312hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.2.0=hfb803bf_1
stack_data=0.6.2=pyhd8ed1ab_0
svt-av1=2.0.0=h63175ca_0
sympy=1.12=py312haa95532_0
tbb=2021.12.0=h91493d7_0
tk=8.6.13=h5226925_1
torchaudio=2.2.2=pypi_0
torchvision=0.17.2=pypi_0
tornado=6.4=py312he70551f_0
tqdm=4.66.2=py312hfc267ef_0
traitlets=5.14.3=pyhd8ed1ab_0
typing_extensions=4.9.0=py312haa95532_1
tzdata=2024a=h04d1e81_0
ucrt=10.0.20348.0=haa95532_0
unicodedata2=15.1.0=py312h2bbff1b_0
urllib3=2.1.0=py312haa95532_1
vc=14.3=hcf57466_18
vc14_runtime=14.38.33130=h82b7239_18
vs2015_runtime=14.38.33130=hcb4865c_18
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.41.2=py312haa95532_0
win_inet_pton=1.1.0=py312haa95532_0
x264=1!164.3095=h8ffe710_2
x265=3.5=h2d74725_3
xorg-libxau=1.0.11=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.4.6=h8cc25b3_0
yaml=0.2.5=he774522_0
zeromq=4.3.5=h63175ca_1
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=hcfcfb64_5
zstd=1.5.5=hd43e919_0
